+++ src/main/java/org/example/data/SetMaps.java
@@ -30,50 +30,26 @@ public class SetMaps {
     protected Map<String, List<Object>> setterBookMap(String title, String author, String isbn,
                                                       int id, int command, Map<String, List<Object>> bookMap) {
         try (Connection connection = DriverManager.getConnection(bibUrl, user, password)) {
-            Statement statement = connection.createStatement();
+            String sql = null;
             List<Object> mapList = new ArrayList<>();
-            int index = id; // Index für das neu erstellte Buch
 
             // Fügt den Titel, Author und ISBN in eine Liste und diese dann in die Buch-Map
             mapList.add(title);
             mapList.add(author);
             mapList.add(isbn);
 
-            // SQL Befehl zum Einfügen von einem neuen Buch
-            String insertCommand = "INSERT INTO book(title, author, isbn) " +
-                    "VALUES('" + title + "', '" + author + "', '" + isbn + "');";
-
-            if (command == 1) {
-                // SQL Befehl zum Überarbeiten vom Buch.
-                insertCommand = "UPDATE book SET " +
-                        "title = '" + title + "' , " +
-                        "author = '" + author + "' , " +
-                        "isbn = '" + isbn + "' " +
-                        "where id = " + id;
-                bookMap.remove(String.valueOf(index));
-                bookMap.put(String.valueOf(index), mapList);
-            } else if (command == 2) {
-                // SQL Befehl zum Löschen von einem Buch
-                insertCommand = "DELETE from book where id = " + id;
-                bookMap.remove(String.valueOf(index));
-            }
-
-            statement.executeUpdate(insertCommand);
-
             if (command == 0) {
-                // Abfrage zum Bestimmen der ID des neuen Buches
-                ResultSet resultSet = statement.executeQuery("SELECT * FROM book");
-
-                while (resultSet.next()) {
-                    index = resultSet.getInt("id");
-                }
-
-                bookMap.put(String.valueOf(index), mapList);
+                sql = "INSERT INTO book (title, author, isbn) VALUES(?, ?, ?) RETURNING id;";
+            } else if (command == 1) {
+                sql = "UPDATE book SET title = ?, author = ?, isbn = ? WHERE id = ?;";
+            } else if (command == 2) {
+                sql = "DELETE FROM book WHERE id = ?;";
             }
 
-            return bookMap;
+            return getTableMap(connection, sql, command, id, title, author, isbn, mapList, bookMap);
         } catch (SQLException e) {
-            throw new RuntimeException(e);
+            System.out.println(e.getMessage());
+            return bookMap;
         }
     }
 
@@ -88,56 +64,98 @@ public class SetMaps {
      *                <ul>
      *                <li>0: Neues Mitglied hinzufügen</li>
      *                <li>1: Mitglied aktualisieren</li>
-     *                <li>0: Mitglied löschen</li>
+     *                <li>2: Mitglied löschen</li>
      *                </ul>
      * */
     protected Map<String, List<Object>> setterMemberMap(String name, String email, String phone, int id,
                                                int command, Map<String, List<Object>> memberMap) {
         try (Connection connection = DriverManager.getConnection(bibUrl, user, password)) {
-            Statement statement = connection.createStatement();
+            String sql = null;
             List<Object> mapList = new ArrayList<>();
-            int index = id; // Index für das neu erstellte Buch
 
             // Fügt den Titel, Author und ISBN in eine Liste und diese dann in die Buch-Map
             mapList.add(name);
             mapList.add(email);
             mapList.add(phone);
 
-            // SQL Befehl zum Einfügen von einem neuen Buch
-            String insertCommand = "INSERT INTO member(name, email, phone) " +
-                    "VALUES('" + name + "', '" + email + "', '" + phone + "');";
-
-            if (command == 1) {
-                // SQL Befehl zum Überarbeiten vom Buch.
-                insertCommand = "UPDATE member SET " +
-                        "name = '" + name + "' , " +
-                        "email = '" + email + "' , " +
-                        "phone = '" + phone + "' " +
-                        "where id = " + id;
-                memberMap.remove(String.valueOf(index));
-                memberMap.put(String.valueOf(index), mapList);
+            if (command == 0) {
+                // SQL Befehl zum Einfügen eines neuen Mitglieds
+                sql =   "INSERT INTO member (name, email, phone) VALUES(?, ?, ?) RETURNING id;";
+            } else if (command == 1) {
+                // SQL Befehl zum Aktualisieren eines Mitglieds
+                sql = "UPDATE member SET name = ?, email = ?, phone = ? WHERE id = ?";
             } else if (command == 2) {
-                // SQL Befehl zum Löschen von einem Buch
-                insertCommand = "DELETE from member where id = " + id;
-                memberMap.remove(String.valueOf(index));
+                // SQL Befehl zum Löschen eines Mitglieds
+                sql = "DELETE FROM member WHERE id = ?";
             }
 
-            statement.executeUpdate(insertCommand);
+            return getTableMap(connection, sql, command, id, name, email, phone, mapList, memberMap);
+        } catch (SQLException e) {
+            System.out.println(e.getMessage());
+            return memberMap;
+        }
+    }
 
-            if (command == 0) {
-                // Abfrage zum Bestimmen der ID des neuen Buches
-                ResultSet resultSet = statement.executeQuery("SELECT * FROM book");
+    /**
+     * In dieser Methode wird der SQL Befehl ausgeführt und die aktualisierte Map wieder ausgegeben.
+     *
+     * @param connection Die Verbindung zum Server.
+     * @param sql Der Befehl für die Datenbank.
+     * @param command Statuscode:
+     *                <ul>
+     *                    <li>0: Hinzufügen</li>
+     *                    <li>1: Aktualisieren</li>
+     *                    <li>2: Löschen</li>
+     *                </ul>
+     * @param id Die ID vom Inhalt.
+     * @param firstRow Inhalt der ersten Spalte.
+     * @param secondRow Inhalt der zweiten Spalte.
+     * @param thirdRow Inhalt der dritten Spalte.
+     * @param mapList Die Liste mit dem Inhalt der Zeile.
+     * @param map Die Map die aktualisiert werden soll.
+     * @return
+     * <ul>
+     *     <li>{@code Map<String, List<Object>>} Gibt das bearbeitete Map aus.</li>
+     *     <lis>{@code null} Falls es keinen Befehl gab.</lis>
+     * </ul>
+     * */
+    private Map<String, List<Object>> getTableMap(Connection connection, String sql, int command, int id,
+                                                  String firstRow, String secondRow, String thirdRow,
+                                                  List<Object> mapList, Map<String, List<Object>> map) throws SQLException {
+        if (sql != null) {
+            try (PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
+                if (command == 0) {
+                    // Fügt den Inhalt in den Befehl zum Einfügen
+                    preparedStatement.setString(1, firstRow);
+                    preparedStatement.setString(2, secondRow);
+                    preparedStatement.setString(3, thirdRow);
+                    ResultSet resultSet = preparedStatement.executeQuery();
 
-                while (resultSet.next()) {
-                    index = resultSet.getInt("id");
+                    // Bestimmt die ID und fügt alles in die Map
+                    if (resultSet.next()) {
+                        int newID = resultSet.getInt("id");
+
+                        map.put(String.valueOf(newID), mapList);
+                    }
+                } else if (command == 1) {
+                    // Fügt den Inhalt in den Befehl zum Aktualisieren
+                    preparedStatement.setString(1, firstRow);
+                    preparedStatement.setString(2, secondRow);
+                    preparedStatement.setString(3, thirdRow);
+                    preparedStatement.setInt(4, id);
+                    preparedStatement.executeUpdate();
+
+                    map.put(String.valueOf(id), mapList);
+                } else if (command == 2) {
+                    // Fügt den Inhalt in den Befehl zum Löschen
+                    preparedStatement.setInt(1, id);
+                    preparedStatement.executeUpdate();
+                    map.remove(String.valueOf(id));
                 }
 
-                memberMap.put(String.valueOf(index), mapList);
+                return map;
             }
-            
-            return memberMap;
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
         }
+        return null;
     }
 }